Restore local/test development mode
While the Container Escape Bounty CTF
profiles are only defined on real VMs,
it is still useful to have a local mode
for testing and development.

This restores the Docker-in-Docker functionality
that was broken when the switch from wss:// to ws://
happened and TLSClientConfig was commented out.
This is done by introducing a new parameter --tlsws.
The SELinux/AppArmor button can be shown when the
OS flag is specified through make run ARGS="-os fedora"
(or ubuntu). The commit also adds the necessary
port forwarding in the Makefile for "make dind".
Finally, the README is updated to reflect that.

Signed-off-by: Kai Lüke <kai@kinvolk.io>